Extracted interface StoredMessageCollection as preparation for implementation based on the new map component and implemented ListBasedStoredMessageCollection in order to support existing implementation.
