+++ greenmail-core/src/main/java/com/icegreen/greenmail/store/InMemoryStore.java
@@ -6,13 +6,10 @@
  */
 package com.icegreen.greenmail.store;
 
-import com.icegreen.greenmail.foedus.util.MsgRangeFilter;
 import com.icegreen.greenmail.imap.ImapConstants;
-import com.icegreen.greenmail.mail.MovingMessage;
 
-import javax.mail.*;
-import javax.mail.internet.MimeMessage;
-import javax.mail.search.SearchTerm;
+import javax.mail.MessagingException;
+import javax.mail.Quota;
 import java.util.*;
 
 /**
@@ -24,18 +21,8 @@ import java.util.*;
  */
 public class InMemoryStore
         implements Store, ImapConstants {
-    private RootFolder rootMailbox = new RootFolder();
-    private static final Flags PERMANENT_FLAGS = new Flags();
     boolean quotaSupported = true;
-
-    static {
-        PERMANENT_FLAGS.add(Flags.Flag.ANSWERED);
-        PERMANENT_FLAGS.add(Flags.Flag.DELETED);
-        PERMANENT_FLAGS.add(Flags.Flag.DRAFT);
-        PERMANENT_FLAGS.add(Flags.Flag.FLAGGED);
-        PERMANENT_FLAGS.add(Flags.Flag.SEEN);
-    }
-
+    private RootFolder rootMailbox = new RootFolder();
     private Map<String, Set<Quota>> quotaMap = new HashMap<String, Set<Quota>>();
 
     public MailFolder getMailbox(String absoluteMailboxName) {
@@ -267,445 +254,6 @@ public class InMemoryStore
             HierarchicalFolder child = (HierarchicalFolder) aChildren;
             mailboxes.add(child);
             addAllChildren(child, mailboxes);
-        }
-    }
-
-    private class RootFolder extends HierarchicalFolder {
-        public RootFolder() {
-            super(null, ImapConstants.USER_NAMESPACE);
-        }
-
-        public String getFullName() {
-            return name;
-        }
-    }
-
-    private class HierarchicalFolder implements MailFolder, UIDFolder {
-        private Collection<HierarchicalFolder> children;
-        private HierarchicalFolder parent;
-
-        protected String name;
-        private boolean isSelectable = false;
-
-        private final List<StoredMessage> mailMessages = Collections.synchronizedList(new ArrayList<StoredMessage>());
-        private long nextUid = 1;
-        private long uidValidity;
-
-        private final List<FolderListener> _mailboxListeners = Collections.synchronizedList(new ArrayList<FolderListener>());
-
-        public HierarchicalFolder(HierarchicalFolder parent,
-                                  String name) {
-            this.name = name;
-            this.children = new ArrayList<HierarchicalFolder>();
-            this.parent = parent;
-            this.uidValidity = System.currentTimeMillis();
-        }
-
-        public Collection<HierarchicalFolder> getChildren() {
-            return children;
-        }
-
-        public HierarchicalFolder getParent() {
-            return parent;
-        }
-
-        public void moveToNewParent(HierarchicalFolder newParent) {
-            if (!newParent.getChildren().contains(this)) {
-                parent = newParent;
-                parent.getChildren().add(this);
-            }
-        }
-
-        public HierarchicalFolder getChild(String name) {
-            for (HierarchicalFolder child : children) {
-                if (child.getName().equalsIgnoreCase(name)) {
-                    return child;
-                }
-            }
-            return null;
-        }
-
-        public void setName(String name) {
-            this.name = name;
-        }
-
-        public String getName() {
-            return name;
-        }
-
-        public String getFullName() {
-            return parent.getFullName() + HIERARCHY_DELIMITER_CHAR + name;
-        }
-
-        public Flags getPermanentFlags() {
-            return PERMANENT_FLAGS;
-        }
-
-        public int getMessageCount() {
-            synchronized (mailMessages) {
-                return mailMessages.size();
-            }
-        }
-
-        public long getUidValidity() {
-            return uidValidity;
-        }
-
-        public long getUidNext() {
-            return nextUid;
-        }
-
-        public int getUnseenCount() {
-            int count = 0;
-            synchronized (mailMessages) {
-                for (StoredMessage message : mailMessages) {
-                    if (!message.isSet(Flags.Flag.SEEN)) {
-                        count++;
-                    }
-                }
-            }
-            return count;
-        }
-
-        /**
-         * Returns the 1-based index of the first unseen message. Unless there are outstanding
-         * expunge responses in the ImapSessionMailbox, this will correspond to the MSN for
-         * the first unseen.
-         */
-        public int getFirstUnseen() {
-            synchronized (mailMessages) {
-                for (int i = 0; i < mailMessages.size(); i++) {
-                    StoredMessage message = mailMessages.get(i);
-                    if (!message.isSet(Flags.Flag.SEEN)) {
-                        return i + 1;
-                    }
-                }
-            }
-            return -1;
-        }
-
-        public int getRecentCount(boolean reset) {
-            int count = 0;
-            synchronized (mailMessages) {
-                for (StoredMessage message : mailMessages) {
-                    if (message.isSet(Flags.Flag.RECENT)) {
-                        count++;
-                        if (reset) {
-                            message.setFlag(Flags.Flag.RECENT, false);
-                        }
-                    }
-                }
-            }
-            return count;
-        }
-
-        public int getMsn(long uid) throws FolderException {
-            synchronized (mailMessages) {
-                for (int i = 0; i < mailMessages.size(); i++) {
-                    StoredMessage message = mailMessages.get(i);
-                    if (message.getUid() == uid) {
-                        return i + 1;
-                    }
-                }
-            }
-            throw new FolderException("No such message.");
-        }
-
-        public void signalDeletion() {
-            // Notify all the listeners of the new message
-            synchronized (_mailboxListeners) {
-                for (FolderListener listener : _mailboxListeners) {
-                    listener.mailboxDeleted();
-                }
-            }
-
-        }
-
-        public List getMessages(MsgRangeFilter range) {
-            List<StoredMessage> ret = new ArrayList<StoredMessage>();
-
-            synchronized (mailMessages) {
-                for (int i = 0; i < mailMessages.size(); i++) {
-                    if (range.includes(i + 1)) {
-                        ret.add(mailMessages.get(i));
-                    }
-                }
-            }
-
-            return ret;
-        }
-
-        public List<StoredMessage> getMessages() {
-            synchronized (mailMessages) {
-                // Return new list since we don't want to give the caller access to the internal list
-                return new ArrayList<StoredMessage>(mailMessages);
-            }
-        }
-
-        public List<StoredMessage> getNonDeletedMessages() {
-            List<StoredMessage> ret = new ArrayList<StoredMessage>();
-
-            synchronized (mailMessages) {
-                for (StoredMessage mailMessage : mailMessages) {
-                    if (!mailMessage.getFlags().contains(Flags.Flag.DELETED)) {
-                        ret.add(mailMessage);
-                    }
-                }
-            }
-
-            return ret;
-        }
-
-        public boolean isSelectable() {
-            return isSelectable;
-        }
-
-        public void setSelectable(boolean selectable) {
-            isSelectable = selectable;
-        }
-
-        public long appendMessage(MimeMessage message,
-                                  Flags flags,
-                                  Date receivedDate) {
-            long uid = nextUid;
-            nextUid++;
-
-            try {
-                message.setFlags(flags, true);
-                message.setFlag(Flags.Flag.RECENT, true);
-            } catch (MessagingException e) {
-                throw new IllegalStateException("Can not set flags", e);
-            }
-            StoredMessage storedMessage = new StoredMessage(message,
-                    receivedDate, uid);
-
-            int newMsn;
-            synchronized (mailMessages) {
-                mailMessages.add(storedMessage);
-                newMsn = mailMessages.size();
-            }
-
-            // Notify all the listeners of the new message
-            synchronized (_mailboxListeners) {
-                for (FolderListener _mailboxListener : _mailboxListeners) {
-                    _mailboxListener.added(newMsn);
-                }
-            }
-
-            return uid;
-        }
-
-        public void setFlags(Flags flags, boolean value, long uid, FolderListener silentListener, boolean addUid) throws FolderException {
-            int msn = getMsn(uid);
-            StoredMessage message = mailMessages.get(msn - 1);
-
-            message.setFlags(flags, value);
-
-            Long uidNotification = null;
-            if (addUid) {
-                uidNotification = uid;
-            }
-            notifyFlagUpdate(msn, message.getFlags(), uidNotification, silentListener);
-        }
-
-        public void replaceFlags(Flags flags, long uid, FolderListener silentListener, boolean addUid) throws FolderException {
-            int msn = getMsn(uid);
-            StoredMessage message = mailMessages.get(msn - 1);
-            message.setFlags(MessageFlags.ALL_FLAGS, false);
-            message.setFlags(flags, true);
-
-            Long uidNotification = null;
-            if (addUid) {
-                uidNotification = uid;
-            }
-            notifyFlagUpdate(msn, message.getFlags(), uidNotification, silentListener);
-        }
-
-        private void notifyFlagUpdate(int msn, Flags flags, Long uidNotification, FolderListener silentListener) {
-            synchronized (_mailboxListeners) {
-                for (FolderListener listener : _mailboxListeners) {
-                    if (listener == silentListener) {
-                        continue;
-                    }
-
-                    listener.flagsUpdated(msn, flags, uidNotification);
-                }
-            }
-        }
-
-        public void deleteAllMessages() {
-            synchronized (mailMessages) {
-                mailMessages.clear();
-            }
-        }
-
-        public void store(MovingMessage mail) throws Exception {
-            store(mail.getMessage());
-        }
-
-
-        public void store(MimeMessage message) throws Exception {
-            Date receivedDate = new Date();
-            Flags flags = new Flags();
-            appendMessage(message, flags, receivedDate);
-        }
-
-        public StoredMessage getMessage(long uid) {
-            synchronized (mailMessages) {
-                for (StoredMessage mailMessage : mailMessages) {
-                    if (mailMessage.getUid() == uid) {
-                        return mailMessage;
-                    }
-                }
-            }
-            return null;
-        }
-
-        public long[] getMessageUids() {
-            synchronized (mailMessages) {
-                long[] uids = new long[mailMessages.size()];
-                for (int i = 0; i < mailMessages.size(); i++) {
-                    StoredMessage message = mailMessages.get(i);
-                    uids[i] = message.getUid();
-                }
-                return uids;
-            }
-        }
-
-        private void deleteMessage(int msn) {
-            synchronized (mailMessages) {
-                mailMessages.remove(msn - 1); // input is 1 based index
-            }
-        }
-
-        public long[] search(SearchTerm searchTerm) {
-            List<StoredMessage> matchedMessages = new ArrayList<StoredMessage>();
-
-            synchronized (mailMessages) {
-                for (StoredMessage mailMessage : mailMessages) {
-                    if (searchTerm.match(mailMessage.getMimeMessage())) {
-                        matchedMessages.add(mailMessage);
-                    }
-                }
-            }
-
-            long[] matchedUids = new long[matchedMessages.size()];
-            for (int i = 0; i < matchedUids.length; i++) {
-                StoredMessage storedMessage = matchedMessages.get(i);
-                long uid = storedMessage.getUid();
-                matchedUids[i] = uid;
-            }
-            return matchedUids;
-        }
-
-        public void copyMessage(long uid, MailFolder toFolder)
-                throws FolderException {
-            StoredMessage originalMessage = getMessage(uid);
-            MimeMessage newMime;
-            try {
-                newMime = new MimeMessage(originalMessage.getMimeMessage());
-            } catch (MessagingException e) {
-                throw new FolderException("Can not copy message " + uid + " to folder " + toFolder, e);
-            }
-
-            toFolder.appendMessage(newMime, originalMessage.getFlags(), originalMessage.getReceivedDate());
-        }
-
-        public void expunge() throws FolderException {
-            synchronized (mailMessages) {
-                for (int i = mailMessages.size() - 1; i >= 0; i--) {
-                    StoredMessage message = mailMessages.get(i);
-                    if (message.isSet(Flags.Flag.DELETED)) {
-                        expungeMessage(i + 1); // MSNs start counting at 1
-                    }
-                }
-            }
-        }
-
-        private void expungeMessage(int msn) {
-            // Notify all the listeners of the pending delete
-            synchronized (_mailboxListeners) {
-                deleteMessage(msn);
-                for (FolderListener expungeListener : _mailboxListeners) {
-                    expungeListener.expunged(msn);
-                }
-            }
-        }
-
-        public void addListener(FolderListener listener) {
-            synchronized (_mailboxListeners) {
-                _mailboxListeners.add(listener);
-            }
-        }
-
-        public void removeListener(FolderListener listener) {
-            synchronized (_mailboxListeners) {
-                _mailboxListeners.remove(listener);
-            }
-        }
-
-        @Override
-        public String toString() {
-            return "HierarchicalFolder{" +
-                    "name='" + name + '\'' +
-                    ", parent=" + parent +
-                    ", isSelectable=" + isSelectable +
-                    '}';
-        }
-
-        @Override
-        public long getUIDValidity() throws MessagingException {
-            return getUidValidity();
-        }
-
-        @Override
-        public Message getMessageByUID(long uid) throws MessagingException {
-            return getMessage(uid).getMimeMessage();
-        }
-
-        @Override
-        public Message[] getMessagesByUID(long start, long end) throws MessagingException {
-            synchronized (mailMessages) {
-                List<Message> messages = new ArrayList<Message>();
-                for (StoredMessage mailMessage : mailMessages) {
-                    final long uid = mailMessage.getUid();
-                    if (uid >= start && uid <= end) {
-                        messages.add(mailMessage.getMimeMessage());
-                    }
-                }
-                return messages.toArray(new Message[messages.size()]);
-            }
-        }
-
-        @Override
-        public Message[] getMessagesByUID(long[] uids) throws MessagingException {
-            synchronized (mailMessages) {
-                List<Message> messages = new ArrayList<Message>(uids.length);
-                Map<Long, StoredMessage> uid2Msg = new HashMap<Long, StoredMessage>(mailMessages.size());
-                for (StoredMessage mailMessage : mailMessages) {
-                    uid2Msg.put(mailMessage.getUid(), mailMessage);
-                }
-                for (long uid : uids) {
-                    final StoredMessage storedMessage = uid2Msg.get(uid);
-                    if (storedMessage != null) {
-                        messages.add(storedMessage.getMimeMessage());
-                    }
-                }
-                return messages.toArray(new Message[messages.size()]);
-            }
-        }
-
-        @Override
-        public long getUID(Message message) throws MessagingException {
-            // Check if we have a message with same object reference ... otherwise, not supported.
-            synchronized (mailMessages) {
-                for (StoredMessage mailMessage : mailMessages) {
-                    if (mailMessage.getMimeMessage() == message) {
-                        return mailMessage.getUid();
-                    }
-                }
-            }
-            throw new IllegalStateException("No match found for " + message);
         }
     }
