+++ greenmail-core/src/main/java/com/icegreen/greenmail/store/RootFolder.java
@@ -1,0 +1,13 @@
+package com.icegreen.greenmail.store;
+
+import com.icegreen.greenmail.imap.ImapConstants;
+
+class RootFolder extends HierarchicalFolder {
+    public RootFolder() {
+        super(null, ImapConstants.USER_NAMESPACE);
+    }
+
+    public String getFullName() {
+        return name;
+    }
+}
